Temos apostas a funcionar! Dealer ja funciona! Agora so falta dar o dinheiro aos vencedores, verificar quando o jogo acaba
Bug: quango o player pede double, fica sempre com blackjack independemente da carta q recebe LOL

+++ blackjack/blackjack/game.h
@@ -11,10 +11,10 @@ class Game{
     Dealer dealer;
     bool gameOver;
     House gameDeck;
-    int minimBet;
+    float minimBet;
     
 public:
-    Game(vector<Player>, int);
+    Game(vector<Player>, float);
     int inGameNumber();
     Round* playRound();
     bool over();
